added authentification & autorization

+++ backend/routes/posts.js
@@ -2,6 +2,9 @@ const express = require('express');
 const postsRouter = express.Router();
 
 const Post = require('../models/post');
+const checkAuth = require('../security/check-auth');
+
+const multer = require('multer');
 
 const MIME_TYPE_MAP = {
   'image/png': 'png',
@@ -9,7 +12,6 @@ const MIME_TYPE_MAP = {
   'image/jpg': 'jpg',
 };
 
-const multer = require('multer');
 
 const storage = multer.diskStorage({
   destination: (req, file, callback) => {
@@ -27,46 +29,56 @@ const storage = multer.diskStorage({
   }
 });
 
-postsRouter.post('', multer({storage: storage}).single('image'), (req, res, next) => {
-  const url = req.protocol + '://' + req.get('host');
-  // const posts = req.body; // ajouter grâce au BodyParcer
-  const post = new Post({
-    title: req.body.title,
-    content: req.body.content,
-    imagePath: url + '/images/' + req.file.filename
+postsRouter.post('',
+  checkAuth,
+  multer({storage: storage}).single('image'), (req, res, next) => {
+    const url = req.protocol + '://' + req.get('host');
+    // const posts = req.body; // ajouter grâce au BodyParcer
+    const post = new Post({
+      title: req.body.title,
+      content: req.body.content,
+      imagePath: url + '/images/' + req.file.filename,
+      creator: req.userData.userId
+    });
+    post.save().then(createdPost => {
+      res.status(201).json({
+        message: 'Posts ajoutés avec succès',
+        post: {
+          id: createdPost._id, // créé pour éviter d'avoir une erreur en cas de delete juste après ajout (car id:null
+          title: createdPost.title,
+          content: createdPost.content,
+          imagePath: createdPost.imagePath
+        }
+      });
+    });
   });
-  post.save().then(createdPost => {
-    res.status(201).json({
-      message: 'Posts ajoutés avec succès',
-      post: {
-        id: createdPost._id, // créé pour éviter d'avoir une erreur en cas de delete juste après ajout (car id:null
-        title: createdPost.title,
-        content: createdPost.content,
-        imagePath: createdPost.imagePath
+
+postsRouter.put('/:id',
+  checkAuth,
+  multer({storage: storage}).single('image'),
+  (req, res, next) => {
+    let imagePath = req.body.imagePath;
+    if (req.file) {
+      const url = req.protocol + '://' + req.get('host');
+      imagePath = url + '/images/' + req.file.filename
+    }
+    const post = new Post({
+      _id: req.body.id, // obligé pour mettre à jour l'enregistrement
+      title: req.body.title,
+      content: req.body.content,
+      imagePath: imagePath,
+      creator: req.userData.userId
+    });
+    Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(result => {
+      if (result.nModified > 0) {
+        res.status(200).json({message: "Post mis à jour avec succès"});
+      } else {
+        res.status(401).json({message: "Modification non autorisée"});
       }
     });
   });
-});
 
-postsRouter.put('/:id', multer({storage: storage}).single('image'), (req, res, next) => {
-  let imagePath = req.body.imagePath;
-  if (req.file) {
-    const url = req.protocol + '://' + req.get('host');
-    imagePath = url + '/images/' + req.file.filename
-  }
-  const post = new Post({
-    _id: req.body.id, // obligé pour mettre à jour l'enregistrement
-    title: req.body.title,
-    content: req.body.content,
-    imagePath: imagePath
-  });
-  Post.updateOne({_id: req.params.id}, post).then(result => {
-    res.status(200).json({
-      message: "Post mis à jour avec succès"
-    })
-  });
-});
-
+// All posts
 postsRouter.get('', (req, res, next) => {
   const pageSize = +req.query.nbpage;
   const currentPage = +req.query.page;
@@ -80,7 +92,7 @@ postsRouter.get('', (req, res, next) => {
   postQuery
     .then(documents => {
       fetchedPosts = documents;
-      return Post.count();
+      return Post.countDocuments();
     })
     .then(count => {
       res.status(200).json({
@@ -91,7 +103,7 @@ postsRouter.get('', (req, res, next) => {
     });
 });
 
-// All posts
+//one post
 postsRouter.get('/:id', (req, res, next) => {
   Post.findById(req.params.id).then(post => {
     if (post) {
@@ -104,17 +116,18 @@ postsRouter.get('/:id', (req, res, next) => {
   })
 });
 
-postsRouter.delete('/:id', (req, res, next) => {
-  Post.deleteOne({
-    _id: req.params.id
-  })
-    .then(result => {
-      console.log(result);
-      res.status(200).json({
-        message: 'Post effacé'
-      })
-    });
+postsRouter.delete('/:id',
+  checkAuth,
+  (req, res, next) => {
+    Post.deleteOne({_id: req.params.id, creator: req.userData.userId})
+      .then(result => {
+        if (result.n > 0) {
+          res.status(200).json({message: 'Post effacé'});
+        } else {
+          res.status(401).json({message: "Suppression non autorisée"});
+        }
+      });
 
-});
+  });
 
 module.exports = postsRouter;
